added admin upload page

+++ backend/server.js
@@ -1,7 +1,6 @@
 const http = require("http");
 const fs = require('fs');
-const qs = require("querystring");
-const url = require("url");
+const formidable = require('formidable');
 const PORT = process.env.PORT || 3000;
 const db = require(__dirname + "/database/db-methods.js");
 
@@ -34,33 +33,63 @@ const server = http.createServer(function(req,res){
                    res.end();
                 })
              }
+             
+             else if(req.url === '/admin'){
+                fs.readFile(__dirname + '/static/admin.html', function (error, html) {
+                    res.writeHeader(200, {"Content-Type": "text/html"});  
+                    res.write(html);  
+                    res.end();  
+                })
+             }
             break;
         case "POST":
-            res.setHeader('Access-Control-Allow-Origin', '*');
-            res.writeHead(200, {'Content-Type':'application/json;charset=UTF-8'});
-            req.on("data", async function (chunk) {
-                let body = JSON.parse(chunk).body
-                let {action, album, file, size} = body;
-                if(action === 'allAlbums'){
-                    readAlbums(req, res);
-                }else if(action ==='specificAlbum'){
-                    readAlbum(req, res, album);
-                }else if(action === 'likeSong'){
-                    let isLiked = await db.find(file, album);
-                    isLiked ? db.delete(file, album) : db.insert(file, album, size);
-                    res.end(JSON.stringify({isLiked: isLiked}));  
-                }else if(action === 'likedSongs'){
-                    let responseObj = await db.findAll();
-                    responseObj.forEach( file => file['isLiked'] = true);
-                    res.end(JSON.stringify(responseObj));
+            if(req.url === '/'){
+                res.setHeader('Access-Control-Allow-Origin', '*');
+                res.writeHead(200, {'Content-Type':'application/json;charset=UTF-8'});
+                req.on("data", async function (chunk) {
+                    let body = JSON.parse(chunk).body
+                    let {action, album, file, size} = body;
+                    if(action === 'allAlbums'){
+                        readAlbums(req, res);
+                    }else if(action ==='specificAlbum'){
+                        readAlbum(req, res, album);
+                    }else if(action === 'likeSong'){
+                        let isLiked = await db.find(file, album);
+                        isLiked ? db.delete(file, album) : db.insert(file, album, size);
+                        res.end(JSON.stringify({isLiked: isLiked}));  
+                    }else if(action === 'likedSongs'){
+                        let resObj = await db.findAll();
+                        resObj.forEach( file => file['isLiked'] = true);
+                        res.end(JSON.stringify(resObj));
+                    }
+                })
+            }else if( req.url === '/upload'){
+                var dir = `${__dirname}/static/mp3/${Math.floor(Math.random()*100000)}/`;
+
+                if (!fs.existsSync(dir)) {
+                fs.mkdirSync(dir);
                 }
-            })
+                let form = formidable({})
+                form.uploadDir = dir;
+                form.multiples = true;
+                form.parse(req, function (err, fields, files) {
+                    console.log(fields);
+                    console.log(files) // tu są zawarte wszystkie potrzebne dane
+                    files.file.forEach((file)=>{
+                        console.log(file.path, " trololo ", file.name)
+                        fs.rename(file.path, dir + file.name, function (err) {
+                            if(err) console.log(err)
+                            console.log("rename ok")
+                         });
+                    })
+                 });
+            }
             break;
     }
 
     function readAlbum( req, res, album ){
         let location = `${__dirname}/static/mp3/${album}`
-        responseObj = {
+        resObj = {
             album: album,
             files: [],
         };
@@ -73,7 +102,7 @@ const server = http.createServer(function(req,res){
                 if(file.indexOf('.mp3') != -1){
                     // Gets file size
                     var stats = fs.statSync(`${location}/${file}`);
-                    responseObj['files'].push({
+                    resObj['files'].push({
                         file: file, 
                         album: album,
                         isLiked: await db.find(file, album),
@@ -81,11 +110,11 @@ const server = http.createServer(function(req,res){
                     })
                 }  
             }
-            res.end(JSON.stringify(responseObj));  
+            res.end(JSON.stringify(resObj));  
         });    
     }
     function readAlbums( req, res ) {
-        let responseObj = {
+        let resObj = {
             dirs: [],
             files: [],
         }
@@ -94,24 +123,24 @@ const server = http.createServer(function(req,res){
                 return console.log(err);
             }
             files.forEach( file => {
-                responseObj['dirs'].push(file)
+                resObj['dirs'].push(file)
             });
-            fs.readdir(`${__dirname}/static/mp3/${responseObj['dirs'][0]}`, async function (err, files) {
+            fs.readdir(`${__dirname}/static/mp3/${resObj['dirs'][0]}`, async function (err, files) {
                 if (err) {
                     return console.log(err);
                 }
                 for(const file of files){
                     if(file.indexOf('.mp3') != -1){
-                        var stats = fs.statSync(`${__dirname}/static/mp3/${responseObj['dirs'][0]}/${file}`);
-                        responseObj['files'].push({
+                        var stats = fs.statSync(`${__dirname}/static/mp3/${resObj['dirs'][0]}/${file}`);
+                        resObj['files'].push({
                             file: file, 
-                            isLiked: await db.find(file, responseObj['dirs'][0]),
-                            album: responseObj['dirs'][0],
+                            isLiked: await db.find(file, resObj['dirs'][0]),
+                            album: resObj['dirs'][0],
                             size: bytesToSize(stats.size)
                         })
                     }
                 }
-                res.end(JSON.stringify(responseObj));  
+                res.end(JSON.stringify(resObj));  
             });        
         });
     }
